// web/src/api/models/VehicleType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Тип транспорта
 */
export enum VehicleType {
    AIRPLANE = 'airplane',
    CATERING = 'catering',
    REFUELING = 'refueling',
    CLEANING = 'cleaning',
    BAGGAGE = 'baggage',
    FOLLOW_ME = 'follow-me',
    CHARGING = 'charging',
    BUS = 'bus',
    RAMP = 'ramp',
}